requirements.txt
alembic==1.13.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.13.0
async-timeout==5.0.1
bcrypt==5.0.0
billiard==4.2.4
celery==5.4.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.7
ecdsa==0.19.2
et_xmlfile==2.0.0
fastapi==0.115.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
idna==3.11
kombu==5.6.2
loguru==0.7.2
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.4.4
openpyxl==3.1.5
packaging==26.0
pandas==2.2.3
passlib==1.7.4
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
pyasn1==0.6.3
pycparser==3.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2026.1.post1
PyYAML==6.0.3
redis==5.0.8
rsa==4.9.1
ruff==0.15.10
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.15.0
tzdata==2026.1
uvicorn==0.30.6
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
win32_setctime==1.2.0
